Refactor visualization operations
Animation operations are now in a separate animate module. Bits that are
shared between plot and animate now reside in plot_helpers module.
plot_data_frame has been resurrected with an added deprecation notice.
