+++ cate/ops/__init__.py
@@ -54,7 +54,8 @@ from .correlation import pearson_correlation_scalar, pearson_correlation
 from .normalize import normalize, adjust_temporal_attrs, adjust_spatial_attrs
 from .io import open_dataset, save_dataset, read_object, write_object, read_text, write_text, read_json, write_json, \
     read_csv, read_geo_data_frame, read_netcdf, write_netcdf3, write_netcdf4
-from .plot import plot_map, plot, animate_map
+from .plot import plot_map, plot, plot_dataframe
+from .animate import animate_map
 from .resampling import resample_2d, downsample_2d, upsample_2d
 from .subset import subset_spatial, subset_temporal, subset_temporal_index
 from .timeseries import tseries_point, tseries_mean
@@ -93,6 +94,8 @@ __all__ = [
     # .plot
     'plot_map',
     'plot',
+    'plot_dataframe',
+    # .animate
     'animate_map',
     # .io
     'open_dataset',
@@ -129,4 +132,8 @@ __all__ = [
     'oni',
     # .outliers
     'detect_outliers',
+    # .data_frame
+    'data_frame_min',
+    'data_frame_max',
+    'data_frame_query',
 ]
